pyproject.toml
[tool.poetry]
name = "openaspen"
version = "0.1.0"
description = "Open-source tree-structured AI agent framework for multi-LLM orchestration with hierarchical RAG"
authors = ["OpenAspen Contributors"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/yourusername/openaspen"
repository = "https://github.com/yourusername/openaspen"
keywords = ["ai", "agents", "llm", "rag", "langchain", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^1.2.0"
langgraph = "^1.0.0"
langchain-openai = "^1.1.0"
langchain-anthropic = "^1.3.0"
chromadb = "^0.6.0"
faiss-cpu = "^1.7.4"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
click = "^8.1.7"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.1"
tiktoken = "^0.8.0"
openai = "^1.10.0"
anthropic = "^0.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.0"
ruff = "^0.1.14"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
openaspen = "openaspen.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=openaspen --cov-report=term-missing"
